+++ src/interface/mesh_data.cpp
@@ -10,6 +10,8 @@
 // license in this material to reproduce, prepare derivative works, distribute copies to
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
+#include <string>
+
 #include "mesh_data.hpp"
 
 #include "mesh/mesh.hpp"
@@ -17,8 +19,9 @@
 namespace parthenon {
 
 template <typename T>
-void MeshData<T>::Initialize(BlockList_t blocks, Mesh *pmesh,
-                             std::optional<int> gmg_level) {
+MeshData<T>::MeshData(const std::string &name, BlockList_t blocks, Mesh *pmesh,
+                      std::optional<int> gmg_level)
+    : stage_name_(name) {
   const int nblocks = blocks.size();
   block_data_.resize(nblocks);
   SetMeshProperties(pmesh);